Addressing issues with '--disable-swig' for rpm7
When giving '--disable-swig' (which is the default option), 'make rpm7' will
complains about a strange command not found. This is because $(SWIG) automake
variable will not be set (well ... we disabled it), and sos.c, sos.py, bwx.c,
and bwx.py, which are generated by swig, have been listed into distribution list
(for 'make dist-gzip'). So, when the automake system try to gather all sources
in the distribution list, it in turns try to generate those files with unset
$(SWIG), resulting in build failure.

This commit addresses the issue by listing the generated files under 'nodist_'
variables.