Partition generation number ignored
The sos partition generation number was being ignored in the code path
that searched for the active partition if the cached partition was found
to be not primary.

This patch adds a cached value of the generation number used to build
the partition list and rebuilds the list if it changes.

Signed-off-by: Tom Tucker <tom@ogc.us>