Normalize required fields in known schema
These changes normalize the required fields for Bluewaters and Trinity
data when used with the job tools.

In addition an interface was added to allow iterating through all
objects in a partition without an index. The ods was extended to support
object iteration that can be continued at a position after a callback
cancels iteration. This is useful for reindexing large partitions
that would otherwise hold a global container lock for a very long period
of time.

Signed-off-by: Tom Tucker <tom@ogc.us>