Add logic to update obj_sz and pg_sz in the ODS in ods_extend
If another process caused an ods_extend on an ODS, the values
cached in the other processes memory become incorrect. While map_new()
updated these values, ods_extend() did not. Symptoms include truncating
the file instead of extending it, and references found in an index
not matching any valid map in the process and assert'ing.

Signed-off-by: Tom Tucker <tom@ogc.us>