Reduce the memory footprint of SOS containers
This change modifies the way the ODS object store are mmapped.
With this change the files that underpin the store are mapped as-needed,
i.e. when and object reference is turned into a pointer into memory.
At that time, a tree of maps that are indexed by offset:size is searched.
If a suitable map exists, it is used and refcounted, if not, a new one
is added to the tree.

An ODS configuration option called ods_map_sz determines how big a map is
by default. Gigantic objects may require their own maps, but otherwise,
this constrains how large the extents are that map the file. If a
file extent is never touched, it is never mapped.

In order to avoid 'thrashing', i.e. mapping, unmapping, mapping, unapping
a map is not freed when the last object that references it goes away,
instead, there is a garbage collection thread that runs every 10
seconds that traverses the map tree and discards maps that have
not been used in the last 10 seconds. The map timestamp is updated
whenever it is used to return a pointer to a file offset.

As part of this change, the majority of the meta-data has been removed
for the obj file. This prevents errant applications from destroying a
database by accidentally corrupting the meta-data. It also much more
efficiently packs objects into these file pages, especially for smaller
objects. All meta-data is now kept in the 'page' file.