Fix race between garbage collection thread and ods_close()
The garbage collection thread runs every 10s to find
and free maps that have not been recently used. There
was a bug where if the gc_thread_fn was processing an
ODS and ods_close() was called by the application, a
crash would occur.

Added an environment variable to tweak the gc_thread_fn
garbage collection interval.