Modify JOIN attributes so they don't consume DB memory
JOIN attributes previously occupied storage in the object data
section. Because JOIN attributes are constructed from the values
of other attributes in the object, this is not strictly necessary.
This patch changes the JOIN attribute implementation such that they
do not consume space in the object data. This also allows for
JOIN attributes to be added and subtracted from an object schema
without changing the on-disk structure of the object itself.