The code does not support upsampling
If the requested interval_ms was much less than the
sample interval, the code copied the same data into
the result at increasing indices.

To correctly support upsampling, intervening samples
would be interpolated, however this logic is not present.
In the interim, this change simply ignores the input
interval_ms if the data is not present.