join value size bug fix
`join_size_fn()` -- the function that calculate the size of the join
value -- was incorrectly calculating the size of the join value. The
effect of incorrect value size resulted in `ENOENT` when removing an
object from the join-attribute index.

This patch removes the calculation, and simply use
`value->data->array.count` as size becuase join value is a byte array.