Do not free idx_class memory in ods_idx_term() destructor
There is an issue with dlopen()/dlclose() and library destructor
invocation, specifically, dependent libraries such as libsos
may have pointers to the idx_class structures and dependening
on the order in which the library destructors are called, it
is possible to have sos containers pointing to idx_class memory
after this destructor is called. For this reason, the class memory
is not freed. Yes, valgrind will complain, but this is an example
of a false positive.