Add a compound key type to support JOIN
The JOIN key types was mapped to MEMCMP which is not sufficient
to support signed types, and in general creates confusion when
defining filter condition keys. This change creates support for
compound keys across a broader range of types and without the
performance issues caused by sos_filter_begin/sos_filter_end
with unspecified JOIN attributes or attribuates that are 0
valued.