Modify sos_schema_add() to expand container as necessary
The sos_schema_add function was not expanding the schema ODS when
it filled up the default initial size. This arbitrarily limited
the number and size of schema.

Also, sos_container_open would open all indices for all schema. This
effectively limited the number of schema to the ulimit file handle
count / 2. Now schema indices are only opened when accessed.