Refactor DSOS object creation using transactions
Object creation must be inside a transaction boundary. When
dsos_transaction_begin() is called all subsequent object
creations are queued. When dsos_transaction_end() is called
the transaction_begin RPC is sent to the server, the object
creations are sent to the server and then the transaction_end RPC is sent.
The server sees BEGIN, OBJ_CREATE, ..., END.