Import partition infrastructure
This set of changes supports importing partitions with different schema
dictionaries into containers for query and analysis. The
sos_part_migrate command is added that will change the schema-uuid in a
partition's objects from one directory to another. This allows for
partitions created at different times and on different machines to be
combined into a single container with a well-defined and consistent
object schema.