Move the JSON logic into its own library
The sos_part.c logic and sos_schema logic will support
importing, exporting and remapping schema UUID. The
template files for this are coded in JSON. This
change moves the json code to it's own library. An
unrelated change is that the ax_lib_readline macro
was updated to work on IBM Power boxes.