Add a schema UUID remapping capability to SOS
This change creates the infrastructure for remapping a
partition's object UUID from one schema template directory
to another. This may be required if partitions from
containers with different schema UUID are imported into
a third container. For this to work, both partitions
need to use a common UUID for a given schema.