Ensure correct ODS index is used for objects
A SOS index consists of many ODS indices;
one for each partition in the container. It
is assumed that the partition from which an
object was allocated is indexed by the ODS
index in the same partition. This change
ensures objects are inserted into, and
deleted from the correct ODS index.