Fix the array handling logic in SqlQuery.next()
A Pandas DataFrame does not support 2-D numpy arrays. This broke
the logic that allocates the Numpy arrays that are used to store
Sos object attribute values. This change creates a numpy array
of objects, each object of which is another numpy array. A DataFrame
sees this as a 1-D array of objects.