Add options to add/remove indices to sos-index
The sos-index command is used to manage SOS indices. The ability
to add and remove indices for a schema and attribute has been
added. The intended use is to add indices and remove indices
as system conditions vary. For example, indices for which the
key includes the job_id may not be useful when there is no
job running on a node. In this case, it may be desirable to
remove this indices to decrease the load on the storage back-end
and the storage footprint for the monitored data.