Add --verify to the sos-part command
This change adds a partition verification feature to sos-part
that allows an administrator to determine if the files in a
partition are valid ODS containers.

This verification is intended to be a fast sanity-check and
does not perform a time-consuming, deep validation of the
consistency of the data in the container.