Fix dsos_transaction_begin() to honor `timeout`
`dsos_transaction_begin()` should be a 'blocking' call if the `timeout`
argument is NULL, or a timed-wait call if `timeout` parameter is given.

REMARK: The 'nogil' patch in the Cython code is to release the Python's
Global Interpreter Lock (GIL) before we call (with a chance of blocking)
the dsos transaction functions. Otherwise, other Python's threads could
not proceed since the blocking-caller was also holding the GIL.