Fix Python Sos.DsosSchema intializer
DsosSchema initializer set its `dcont` using the supplied `cont`
from the `__init__()` function argument, but then reset `dcont` to
`None`. This seems to be a typological error. With `dcont` being `None`,
the subsequent attributes obtain from DsosSchema will not be able to
perform `find` because it does not have a reference to the DsosContainer.