Add support for explicit names
An explicit name allows the parser to force a quoted
value to a name. This supports the use case where
attribute names contain special characters
like '+', '-', '*', etc.

An example of the syntax follows:

select N'Mem-Avail+able' from meminfo_tom

The N, followed by a single or double quote forces
the quoted value to be interpretted as a name.