Remove fprintf from main in svcapi_svc.c
The svcapi_svc.c file is generated by rpcgen. The
generated code includes an fprintf of a message when
the program exits. The dsosd spins up a new daemon
for each client that attaches. This change removes
that fprintf pollutes the logs with "svc_run returned"
messages every time a client exited.