Fix configure CFLAGS modification in libtirpc
`${CFLAGS}` on the right-hand-side in `Makefile.am` referred to the
environment variable or CLI option to `make` command (e.g.
`CFLAGS=bla make` or `make CFLAGS=bla`). The right-hand-side `${CFLAGS}`
does NOT refer to CFLAGS option to `./configure` script. Hence the
statement:
	CFLAGS := $(filter-out -Werror, ${CFLAGS})
simply ignored the CFLAGS supplied to `./configure` script.

This patch modify the RHS `${CFLAGS}` to `@CFLAGS@` as follows
	CFLAGS := $(filter-out -Werror, @CFLAGS@)
so that the final CFLAGS in this file is configure CFLAGS without
the `-Werror` term.